Scripts: Changed the imgtool.py finder to use 'west topdir'
This should be a little safer than the arbitrary directory hopping we were doing
before.

This will also use Python's __pycache__ directory in our scripts rather than
making a .cache directory manually.

This will also now clue in the user as to why the first invocation is taking a
bit longer to finish.