Nano Command: Removed manually setting up 'west' imports
Since this no longer lives as a probably-not-installed Python package in the SDK
workspace, we don't need to manually set up package imports to avoid the 'west'
tool using us from a different directory. Now, we can assume we're installed as
a proper package, and thus can work as-is with 'west'.